move components inside app folder
