+++ src/App/components/FormQuestion.tsx
@@ -1,7 +1,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faWarning } from '@fortawesome/free-solid-svg-icons';
 import { UseFormRegister, FieldErrors } from 'react-hook-form';
-import type { Inputs } from '../App/Features/Dogs/FormCreateDog';
+import type { Inputs } from '../Features/Dogs/FormCreateDog';
 
 interface ValidationOptions {
   required?: boolean;
