+++ src/main.tsx
@@ -4,7 +4,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App.tsx';
 import DogList from './App/Features/Dogs/DogList.tsx';
 import DogDetail from './App/Features/Dogs/DogDetail.tsx';
-import ErrorPage from './components/ErrorPage.tsx';
+import ErrorPage from './App/components/ErrorPage.tsx';
 import './index.css';
 import { Authenticator } from '@aws-amplify/ui-react';
 import { Amplify } from 'aws-amplify';
